chore(tests): move test to new test file Signed-off-by: Vladimir Lazar <vlazar@redhat.com>
