+++ tests/resources/primary-podify-demo.yaml
@@ -25,14 +25,3 @@ spec:
       hostPort: 5000
     securityContext:
       runAsNonRoot: true
-  - env:
-    - name: TERM
-      value: xterm
-    - name: HOME
-      value: /home/frontend
-    - name: HOSTNAME
-      value: 4daf7345fc17
-    image: quay.io/podman-desktop-demo/podify-demo-frontend:v1
-    name: frontend-podified
-    securityContext:
-      runAsNonRoot: true
